requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
openai==1.58.1
faiss-cpu==1.8.0.post1
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
